td-shim: separate definition and implementation of E820
Move the definition of E820Entry and E820Type out of the td-shim binary.
